third_party/blink/web_tests/external/wpt/service-workers/service-worker/same-site-cookies.https-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] cross-site, form post with no service worker
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
[FAIL] cross-site, form post with fallback
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
[FAIL] cross-site, form post with passthrough
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
[FAIL] same-origin, form post with no service worker and cross-site redirect
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
[FAIL] same-origin, form post with fallback and cross-site redirect
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
[FAIL] same-origin, form post with passthrough and cross-site redirect
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
[FAIL] same-origin, form post with no service worker, cross-site redirect, and same-origin redirect
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
[FAIL] same-origin, form post with fallback, cross-site redirect, and same-origin redirect
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
[FAIL] same-origin, form post with passthrough, cross-site redirect, and same-origin redirect
  assert_not_equals: Unspecified-SameSite cookies are not sent with cross-site requests. got disallowed value "COOKIE_VALUE"
Harness: the test ran to completion.
